更新

+++ src/main/java/com/example/demo/entity/User.java
@@ -1,18 +1,17 @@
 package com.example.demo.entity;
 
 public class User {
-    private long id;
+    private Integer id;
     private String name;
 
 
-    public long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
-
 
     public String getName() {
         return name;
@@ -21,6 +20,5 @@ public class User {
     public void setName(String name) {
         this.name = name;
     }
-
 }
